Deploiment du Model ML

+++ backend_Flask2/src/entities/grade.py
@@ -1,5 +1,5 @@
 
-from sqlalchemy import Column, String, ForeignKey, Integer
+from sqlalchemy import Column, String, ForeignKey, Integer , Float
 from marshmallow import Schema, fields
 from .entity import Entity, Base
 from sqlalchemy.orm import relationship
@@ -16,7 +16,7 @@ class Grade(Entity, Base):
     id = Column(Integer, primary_key=True)
     answer_id = Column(Integer, ForeignKey('answers.id'))
     student_id = Column(Integer, ForeignKey('students.id'))
-    score = Column(Integer)
+    score = Column(Float)
 
     def __init__(self, score, answer_id, student_id):
         Entity.__init__(self)
@@ -27,7 +27,7 @@ class Grade(Entity, Base):
 class GradeSchema(Schema):
     id = fields.Integer()
     student_id= fields.Integer()
-    score=fields.Integer()
+    score=fields.Float()
     answer_id=fields.Integer()
     created_at = fields.DateTime()
     updated_at = fields.DateTime()
